+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This format is based on [Keep a Changelog] and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- `ByteSet::maybe_contains`: A [Bloom filter] function that returns `true` if a
+  byte is maybe contained in `self`, with 1/8 (12.5%) probability.
+
 ## 0.1.0 - 2020-06-01
 
 Initial release!
@@ -19,5 +24,7 @@ Initial release!
 
 [Keep a Changelog]:    http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
+
+[Bloom filter]: https://en.wikipedia.org/wiki/Bloom_filter
 
 [Unreleased]: https://github.com/nvzqz/static-assertions-rs/compare/v0.1.0...HEAD